Update Python version requirements and enhance dependency management
- Updated the required Python version in `pyproject.toml` and `uv.lock` to support versions from 3.10 to 3.14.
- Added new dependencies including `litestar[jinja]`, `ty`, and `ruff` for improved functionality and code quality.
- Updated existing dependencies to their latest versions, including `anyio` and `backrefs`.
- Enhanced the GitHub Actions workflows to support multiple Python versions and improved caching strategies.
- Updated documentation to reflect the new Python version support and installation instructions.

+++ src/pytaskflow/storage/base.py
@@ -27,7 +27,7 @@ class JobStorage(ABC):
 
     @abstractmethod
     def dequeue(
-        self, queues: List[str], timeout_seconds: int, server_id: str, worker_id: str
+        self, queues: List[str], timeout_seconds: float, server_id: str, worker_id: str
     ) -> Optional[Job]: ...
 
     @abstractmethod
